Init IN 2020/10/28

+++ optimize_gemm/MMult_1x4_3.h
@@ -6,12 +6,12 @@
 
 /* Routine for computing C = A * B + C */
 
-void AddDot( int, double *, int, double *, double * );
-void AddDot1x4( int, double *, int,  double *, int, double *, int );
+void AddDot( int, float *, int, float *, float * );
+void AddDot1x4( int, float *, int,  float *, int, float *, int );
 
-void MY_MMult( int m, int n, int k, double *a, int lda, 
-                                    double *b, int ldb,
-                                    double *c, int ldc )
+void MY_MMult( int m, int n, int k, float *a, int lda, 
+                                    float *b, int ldb,
+                                    float *c, int ldc )
 {
   int i, j;
 
@@ -26,7 +26,7 @@ void MY_MMult( int m, int n, int k, double *a, int lda,
 }
 
 
-void AddDot1x4( int k, double *a, int lda,  double *b, int ldb, double *c, int ldc )
+void AddDot1x4( int k, float *a, int lda,  float *b, int ldb, float *c, int ldc )
 {
   /* So, this routine computes four elements of C: 
            C( 0, 0 ), C( 0, 1 ), C( 0, 2 ), C( 0, 3 ).  
@@ -47,7 +47,7 @@ void AddDot1x4( int k, double *a, int lda,  double *b, int ldb, double *c, int l
 
 #define X(i) x[ (i)*incx ]
 
-void AddDot( int k, double *x, int incx,  double *y, double *gamma )
+void AddDot( int k, float *x, int incx,  float *y, float *gamma )
 {
   /* compute gamma := x' * y + gamma with vectors x and y of length n.
      Here x starts at location x with increment (stride) incx and y starts at location y and has (implicit) stride of 1.
