+++ optimize_gemm/MMult_1x4_4.h
@@ -6,11 +6,11 @@
 
 /* Routine for computing C = A * B + C */
 
-void AddDot1x4( int, double *, int,  double *, int, double *, int );
+void AddDot1x4( int, float *, int,  float *, int, float *, int );
 
-void MY_MMult( int m, int n, int k, double *a, int lda, 
-                                    double *b, int ldb,
-                                    double *c, int ldc )
+void MY_MMult( int m, int n, int k, float *a, int lda, 
+                                    float *b, int ldb,
+                                    float *c, int ldc )
 {
   int i, j;
 
@@ -25,7 +25,7 @@ void MY_MMult( int m, int n, int k, double *a, int lda,
 }
 
 
-void AddDot1x4( int k, double *a, int lda,  double *b, int ldb, double *c, int ldc )
+void AddDot1x4( int k, float *a, int lda,  float *b, int ldb, float *c, int ldc )
 {
   /* So, this routine computes four elements of C: 
            C( 0, 0 ), C( 0, 1 ), C( 0, 2 ), C( 0, 3 ).
